Ensure strings are no longer quoted in the object model
if you have (description eq 'red') the Predicate model does not need to
have the value as 'red'.  Instead it should be red, since value is a
member and the object IS a string.  That can lead to double escaping in
some processors and rather than fix downstream code it makes sense to
fix the input